Move console_factory() into utils
