+++ tests/cli/test_oas.py
@@ -1,4 +1,3 @@
-import os
 import tempfile
 from pathlib import Path
 from typing import Any
@@ -9,7 +8,6 @@ import pytest
 import typer
 
 from openapi_spec_tools.cli.oas import DisplayOption
-from openapi_spec_tools.cli.oas import console_factory
 from openapi_spec_tools.cli.oas import content_type_list
 from openapi_spec_tools.cli.oas import diff
 from openapi_spec_tools.cli.oas import info
@@ -36,25 +34,6 @@ from tests.helpers import asset_filename
 PET_YAML = asset_filename("pet.yaml")
 PET2_YAML = asset_filename("pet2.yaml")
 PET3_YAML = asset_filename("pet3.yaml")
-
-
-#################################################
-# Utilities
-
-def test_console_factory() -> None:
-    # when running the tests, the PYTEST_VERSION is defined by default
-    console = console_factory()
-    assert 3000 == console.width
-
-    with mock.patch.dict(os.environ, {"TERMINAL_WIDTH": "12"}):
-        console = console_factory()
-        assert 12 == console.width
-
-    # using default width for the platform, so it seems to vary
-    with mock.patch.dict(os.environ, {}, clear=True):
-        console = console_factory()
-        assert console.width != 3000
-
 
 
 #################################################
